# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Clean Architecture Knowledge API
  version: 0.1.0
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3000
security: []
paths:
...
  /knowledge/search:
    get:
      summary: Search knowledge articles
      operationId: searchKnowledge
      security: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: tag
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Knowledge article list.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/knowledge'
        '400':
          $ref: '#/components/responses/validation-error'
...
components:
  schemas:
...
    knowledge:
      type: object
      required:
        - id
        - title
        - body
        - excerpt
        - tags
        - category
        - path
        - order
        - sourceUrl
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 120
        body:
          type: string
          minLength: 1
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        path:
          type: string
        order:
          type: number
        sourceUrl:
          type: string
        updatedAt:
          type: string
          format: date-time
    error-response:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: Validator-specific error details.
...
  responses:
...
    validation-error:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error-response'
...